auth fragment added.
