+++ app/src/main/AndroidManifest.xml
@@ -13,10 +13,10 @@
         android:theme="@style/Theme.FitAI"
         tools:targetApi="31">
         <activity
-            android:name=".activities.MainActivity"
+            android:name=".MainActivity"
             android:exported="false" />
         <activity
-            android:name=".activities.SplashActivity"
+            android:name=".splash.SplashActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
