+++ app/src/main/java/com/mobven/fitai/MainActivity.kt
@@ -1,11 +1,10 @@
-package com.mobven.fitai.activities
+package com.mobven.fitai
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
-import com.mobven.fitai.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
