+++ app/src/main/res/drawable/dark_background.xml
@@ -1,0 +1,4 @@
+<shape android:shape="rectangle" xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@color/darkButtonBackground"/>
+    <corners android:radius="8dp"/>
+</shape>
