+++ app/src/main/res/drawable/fitai.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="100dp"
+    android:height="37dp"
+    android:viewportWidth="100"
+    android:viewportHeight="37">
+  <path
+      android:pathData="M19.62,3.64V8.1H5.46V17.94H16.5V22.41H5.46V37H-0.01V3.64H19.62ZM27.48,7.05C26.49,7.05 25.65,6.71 24.98,6.04C24.31,5.37 23.97,4.54 23.97,3.54C23.97,2.55 24.31,1.72 24.98,1.05C25.65,0.38 26.49,0.04 27.48,0.04C28.44,0.04 29.25,0.38 29.93,1.05C30.6,1.72 30.93,2.55 30.93,3.54C30.93,4.54 30.6,5.37 29.93,6.04C29.25,6.71 28.44,7.05 27.48,7.05ZM30.17,10.55V37H24.69V10.55H30.17ZM43.73,15.02V29.66C43.73,30.65 43.96,31.37 44.41,31.82C44.89,32.23 45.69,32.44 46.81,32.44H50.17V37H45.85C43.38,37 41.49,36.42 40.18,35.27C38.87,34.12 38.21,32.25 38.21,29.66V15.02H35.09V10.55H38.21V3.98H43.73V10.55H50.17V15.02H43.73ZM75.71,31.48H63.71L61.88,37H51.99L64.33,3.16H75.18L87.47,37H77.53L75.71,31.48ZM73.35,24.28L69.71,13.34L66.11,24.28H73.35ZM99.94,3.16V37H90.53V3.16H99.94Z"
+      android:fillColor="#000000"/>
+</vector>
