+++ app/src/main/res/drawable/white_background.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape android:shape="rectangle" xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="#F5F8FA"/>
+    <stroke
+        android:width="1dp"
+    android:color="@color/darkButtonBackground"/>
+    <corners android:radius="8dp"/>
+</shape>
