+++ app/src/main/res/layout/activity_splash.xml
@@ -1,11 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.SplashActivity">
+    tools:context=".splash.SplashActivity">
 
     <ImageView
         android:id="@+id/logo"
@@ -15,23 +16,18 @@
         android:src="@drawable/ic_logo"
         app:layout_constraintEnd_toEndOf="@id/guidelineVerticalRight"
         app:layout_constraintStart_toStartOf="@id/guidelineVerticalLeft"
-        app:layout_constraintTop_toTopOf="@id/guidelineHorizontalTop"
-        tools:layout_editor_absoluteX="165dp"
-        tools:layout_editor_absoluteY="309dp" />
+        app:layout_constraintTop_toTopOf="@id/guidelineHorizontalTop"/>
 
-    <TextView
-        android:id="@+id/appName"
+    <ImageView
+        android:id="@+id/name"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:fontFamily="@font/poppins_regular"
-        android:text="@string/app_name"
-        android:textColor="@color/black"
-        android:textSize="30sp"
-        app:layout_constraintBottom_toBottomOf="@id/guidelineHorizontalBottom"
+        android:contentDescription="@string/contentDescription"
+        android:src="@drawable/fitai"
         app:layout_constraintEnd_toEndOf="@id/guidelineVerticalRight"
         app:layout_constraintStart_toStartOf="@id/guidelineVerticalLeft"
         app:layout_constraintTop_toBottomOf="@id/logo"
-        android:layout_marginBottom="50dp"/>
+        />
