+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,6 @@
 <resources>
     <string name="app_name">FitAI</string>
     <string name="contentDescription">Image</string>
+    <string name="register">Kayıt Ol</string>
+    <string name="login">Giriş Yap</string>
 </resources>
